backend/requirements.txt
# --- Web Framework ---
fastapi
uvicorn[standard]
python-dotenv

# --- Data & ML ---
numpy
pandas
scikit-learn
xgboost
joblib
pillow
httpx
opencv-python
statsmodels

# --- Database & ORM ---
sqlalchemy
sqlmodel
alembic
psycopg2-binary
python-multipart

# --- Security & Auth ---
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]

# --- Email Validation ---
pydantic[email]
email-validator

# --- Roboflow Integration ---
inference-sdk

# --- Kaggle Integration ---
kaggle>=1.6.0
